Update README.md to note the client secret JWT expiry
The post from Okta explains how to configure Sign In with Apple and get it working locally.

The issue is that the statically generated JWT token has a maximum lifetime of 6 months (a rule from Apple). I recently wrote about an approach off the back of the Okta article on generating the client secret for each request to ensure you never face an issue with the client secret expiring.